trans/src/requirements.txt
whispercpp
ffmpeg-python
uvicorn
fastapi
numpy
google-cloud-storage
pathlib